feat: code review align, add techie font to tailwind config
