--- src/components/shared/CopyableAddressV2/CopyableAddressV2.css
@@ -1,3 +1,0 @@
-.address {
-  font-weight: normal;
-}
