--- src/components/shared/CopyableAddressV2/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CopyableAddressV2';
