--- src/components/shared/MaskedAddress/MaskedAddress.css
@@ -1,10 +1,0 @@
-.address {
-  font-family: var(--family-techie);
-  font-size: var(--size-smallish);
-  font-weight: 600;
-  line-height: var(--size-smallish);
-}
-
-.middleSection {
-  margin: 0 5px;
-}
