+++ src/components/v5/shared/CopyableAddress/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CopyableAddress';
