+++ src/components/v5/shared/UserAvatarDetails/UserAvatarDetails.tsx
@@ -2,12 +2,12 @@ import React, { FC } from 'react';
 import clsx from 'clsx';
 
 import Avatar from '~v5/shared/Avatar';
+import CopyableAddress from '~v5/shared/CopyableAddress';
 import Icon from '~shared/Icon';
 import { UserAvatarDetailsProps } from './types';
 import UserStatus from '~v5/common/Pills/UserStatus';
 import { UserStatusMode } from '~v5/common/Pills/types';
 import { formatText } from '~utils/intl';
-import CopyableAddressV2 from '~shared/CopyableAddressV2';
 
 const displayName = 'v5.UserAvatarDetails';
 
@@ -74,9 +74,7 @@ const UserAvatarDetails: FC<UserAvatarDetailsProps> = ({
           )}
         </div>
         <p className="text-sm text-gray-600">
-          {walletAddress && (
-            <CopyableAddressV2>{walletAddress}</CopyableAddressV2>
-          )}
+          {walletAddress && <CopyableAddress address={walletAddress} />}
         </p>
       </div>
     </div>
